### 2021-02-04 #3004 @whedon
[PRE REVIEW]: gmr: Gaussian Mixture Regression
**Submitting author:** @AlexanderFabisch (<a href="http://orcid.org/0000-0003-2824-7956">Alexander Fabisch</a>)
**Repository:** <a href="https://github.com/AlexanderFabisch/gmr" target ="_blank">https://github.com/AlexanderFabisch/gmr</a>
**Version:** 1.4
