### 2021-02-05 #3004 @mikldk
@AlexanderFabisch - thanks for this submission.

Can you please maybe comment on how your contribution relates to the `mclust` package in `R`, see e.g. <https://mclust-org.github.io/mclust/> where also a few publications are mentioned, including Fraley C. and Raftery A. E. (2002) *Model-based clustering, discriminant analysis and density estimation*, Journal of the American Statistical Association, 97/458, pp. 611-631.

Do you have any suggestions for reviewers?
