### 2021-02-05 #3004 @AlexanderFabisch
Hi @mikldk,

if I didn't miss anything, mclust focuses on clustering and density estimation, while gmr focuses on multimodal regression. mclust is an R package and gmr is a Python package.

I am sorry, I don't have any suggestion for a reviewer, requirements would be knowledge about Python and machine learning or something similar.
