### 2021-02-22 #3054 @whedon
[REVIEW]: gmr: Gaussian Mixture Regression
**Submitting author:** @AlexanderFabisch (<a href="http://orcid.org/0000-0003-2824-7956">Alexander Fabisch</a>)
**Repository:** <a href="https://github.com/AlexanderFabisch/gmr" target ="_blank">https://github.com/AlexanderFabisch/gmr</a>
**Version:** v1.6.1
