### 2021-03-06 AlexanderFabisch/gmr#25 @vs74
Thanks, @inakleinbottle for creating an issue. I was also going to open an issue that had some common points that you mentioned, but now I guess there won't be a need for it. I am referencing this issue to the [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/3054).
