### 2021-03-06 3f7ddff
Remove duplicate Figure
