### 2021-03-06 cd9de2a
Don't use deprecated numpy types
