### 2021-03-06 5ee4ded
Don't use deprecated numpy types
