### 2021-03-06 63b6d47
Fix type in Gaussian
