+++ README.rst
@@ -50,15 +50,26 @@ You can also install `gmr` from source:
 Example
 -------
 
-Estimate GMM from samples and sample from GMM:
+Estimate GMM from samples, sample from GMM, and make predictions:
 
 .. code-block:: python
 
+    import numpy as np
     from gmr import GMM
 
-    gmm = GMM(n_components=3, random_state=random_state)
+    # Your dataset as a NumPy array of shape (n_samples, n_features):
+    X = np.random.randn(100, 2)
+
+    gmm = GMM(n_components=3, random_state=0)
     gmm.from_samples(X)
+
+    # Estimate GMM with expectation maximization:
     X_sampled = gmm.sample(100)
+
+    # Make predictions with known values for the first feature:
+    x1 = np.random.randn(20, 1)
+    x1_index = [0]
+    x2_predicted_mean = gmm.predict(x1_index, x1)
 
 
 For more details, see:
