### 2021-03-06 c7d822a
Add contribution instructions and meta info
