### 2021-03-06 5dfc4ba
Add API documentation to readme
