### 2021-03-06 47d4a5c
Add API documentation to readme
