+++ doc/apidoc.md
@@ -1,0 +1,561 @@
+Module gmr
+==========
+gmr
+===
+
+Gaussian Mixture Models (GMMs) for clustering and regression in Python.
+
+Sub-modules
+-----------
+* gmr.gmm
+* gmr.mvn
+* gmr.tests
+* gmr.utils
+
+Functions
+---------
+
+    
+`covariance_initialization(X, n_components)`
+:   Initialize covariances.
+    
+    The standard deviation in each dimension is set to the average Euclidean
+    distance of the training samples divided by the number of components.
+    
+    Parameters
+    ----------
+    X : array-like, shape (n_samples, n_features)
+        Samples from the true distribution.
+    
+    n_components : int (> 0)
+        Number of MVNs that compose the GMM.
+    
+    Returns
+    -------
+    initial_covariances : array, shape (n_components, n_features, n_features)
+        Initial covariances
+
+    
+`kmeansplusplus_initialization(X, n_components, random_state=None)`
+:   k-means++ initialization for centers of a GMM.
+    
+    Initialization of GMM centers before expectation maximization (EM).
+    The first center is selected uniformly random. Subsequent centers are
+    sampled from the data with probability proportional to the squared
+    distance to the closest center.
+    
+    Parameters
+    ----------
+    X : array-like, shape (n_samples, n_features)
+        Samples from the true distribution.
+    
+    n_components : int (> 0)
+        Number of MVNs that compose the GMM.
+    
+    random_state : int or RandomState, optional (default: global random state)
+        If an integer is given, it fixes the seed. Defaults to the global numpy
+        random number generator.
+    
+    Returns
+    -------
+    initial_means : array, shape (n_components, n_features)
+        Initial means
+
+    
+`plot_error_ellipse(ax, mvn, color=None, alpha=0.25, factors=array([0.25, 0.5 , 0.75, 1.  , 1.25, 1.5 , 1.75, 2.  ]))`
+:   Plot error ellipse of MVN.
+    
+    Parameters
+    ----------
+    ax : axis
+        Matplotlib axis.
+    
+    mvn : MVN
+        Multivariate normal distribution.
+    
+    color : str, optional (default: None)
+        Color in which the ellipse should be plotted
+    
+    alpha : int, optional (default: 0.25)
+        Alpha value for ellipse
+    
+    factors : array, optional (default: np.linspace(0.25, 2.0, 8))
+        Multiples of the standard deviations that should be plotted.
+
+    
+`plot_error_ellipses(ax, gmm, colors=None, alpha=0.25, factors=array([0.25, 0.5 , 0.75, 1.  , 1.25, 1.5 , 1.75, 2.  ]))`
+:   Plot error ellipses of GMM components.
+    
+    Parameters
+    ----------
+    ax : axis
+        Matplotlib axis.
+    
+    gmm : GMM
+        Gaussian mixture model.
+    
+    colors : list of str, optional (default: None)
+        Colors in which the ellipses should be plotted
+    
+    alpha : int, optional (default: 0.25)
+        Alpha value for ellipses
+    
+    factors : array, optional (default: np.linspace(0.25, 2.0, 8))
+        Multiples of the standard deviations that should be plotted.
+
+Classes
+-------
+
+`GMM(n_components, priors=None, means=None, covariances=None, verbose=0, random_state=None)`
+:   Gaussian Mixture Model.
+    
+    Parameters
+    ----------
+    n_components : int
+        Number of MVNs that compose the GMM.
+    
+    priors : array, shape (n_components,), optional
+        Weights of the components.
+    
+    means : array, shape (n_components, n_features), optional
+        Means of the components.
+    
+    covariances : array, shape (n_components, n_features, n_features), optional
+        Covariances of the components.
+    
[…]
