### 2021-03-06 d6c2f6d
Explain that diagonal covariances are a bad fit
