### 2021-03-06 AlexanderFabisch/gmr#25 @AlexanderFabisch
Thanks for the review. That's really helpful for me. I add my changes to #26 and I will comment on your comments here in the next days.
