### 2021-03-06 AlexanderFabisch/gmr#25 @AlexanderFabisch
> I'm not sure how, but my first run of the test failed. Reinstalling using pip install -e . seemed to fix this. It might be an error from my virtual environment, but you might want to look into this.

The problem is that `GMM.extract_mvn` is not yet available in version 1.4 at PyPI. I will make a new release in the next days, Sorry for the inconveniences.
