+++ paper/paper.bib
@@ -64,3 +64,16 @@ volume={12},
 pages={2825--2830},
 year={2011}
 }
+
+@incollection{Billard2008,
+author="Billard, Aude and Calinon, Sylvain and Dillmann, R{\"u}diger and Schaal, Stefan",
+title="Robot Programming by Demonstration",
+editor="Siciliano, Bruno and Khatib, Oussama",
+booktitle="Springer Handbook of Robotics",
+year="2008",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="1371--1394",
+isbn="978-3-540-30301-5",
+doi="10.1007/978-3-540-30301-5_60"
+}
