### 2021-03-07 AlexanderFabisch/gmr#25 @AlexanderFabisch
> Substantial scholarly effort: I have no problem with the contribution you have made to the field and the functionality of the package. However, I feel you need to stress the differences between your package and the functionality available from Scikit-Learn in the paper. Indeed, in the paper you state "The prediction process for regression is not available in scikit-learn and, thus, will be provided by gmr." To what regression are you referring? The GaussianMixture class has a predict method which does, by my reckoning, perform some kind of prediction. Crucially, this "prediction" is different from the "prediction" provided by the GMM class in gmr. In the paper, you should elaborate on the difference between regression processes used here and clearly state what is in your package but not in Scikit-Learn or other packages.

In #26, I added the following lines:

Note that while sklearn's the function GaussianMixture.predict exists,
it does not perform regression. This function computes the index of the
Gaussian for which the inputs have the highest probability. Furthermore,
multi-modal regression often requires a more complicated interface to extract
not just the mean but also individual Gaussians or sample from the predicted
distribution.

> # Functionality
> Installation. My first attempt to install the package failed because the dependencies are incorrectly set setup.py. Dependencies that are required for the installation and running of your package should be listed under the `install_requires` keyword argument of `setup` rather than `requires`. I think you need 2 dependencies for this package: numpy and scipy. You might also have a dependency on matplotlib, since a number of functions require matplotlib. Scikit-learn is also required for some examples.

This should be fixed with #26, which
* Changes requires to install_requires in setup.py
* Adds optional requirements for examples, tests, and documentation

> Functionality. The package appears to perform as claimed, though I have yet to test an example of my construction. I will make a follow-up post on this issue if I cannot get this to work. The code generally looks to be well written and you have plenty of examples to illustrate the functionality. However, one run of your test suite failed and I'm not sure why. Reinstalling gmr with the local package seemed to fix the issue (pip install -e .). You might want to look into this. Test output below. The second run passed without issue.

See my last comment.

> # Documentation
>
> Statement of need. Your statement of need indicates the problems that your package is designed to solve, but could be a bit more clear about who the audience is. This need not be much, just a sentence to give some suggestions of what kind of people might use this software.

I don't know if I understood correctly what you mean by "audience". I added the following paragraph:

Multimodal regression and Gaussian mixture regression has been used mostly
by the robotics community. For example, inverse problems such as inverse
kinematics cannot be easily modeled with standard regression approaches.
Furthermore, Gaussian mixture regression is the basis of many programming
by demonstration approaches [@Billard2008].

> Example usage. You give lots of examples, but they all seem a bit manufactured. There is a version of the Iris classification problem from Scikit-Learn, which is a great example, but is not explained and it doesn't appear to do a particularly good job at modelling the data.

I added this to the docstring of `plot_iris_from_sklearn.py`:

This is just for demonstration purposes and does not represent an example
of a particularly good fit. Take a look at `plot_iris.py` for a fit with
full covariances.

Many examples are a bit simplistic. It is actually quite difficult to find good, publicly available "real" datasets to demonstrate the library with a more advanced example. I can maybe try to push one of my colleagues to publish his dataset of bimanual robotic manipulation trajectories, which demonstrates quite nicely how we can use GMR, but this might take some time.

> Functionality documentation. You need to generate API documentation for your package. All of the functions have documentation strings in the code, but users probably don't want to have to read through the code to find the functions they need. You should compile some more extensive documentation that covers the functions available to the end user on each class, and illustrate their use.

Thanks for pushing me in this direction. I didn't want to invest too much effort in configuring a complicated sphinx documentation. I found a nice tool with almost no configuration overhead: pdoc3. The API documentation is now available here: https://alexanderfabisch.github.io/gmr/

> Community guidelines. You need to add some guidance for people who want to contribute, report issues, or get support on your package. This can be as simple as pointing them to the relevant features of Github (issues, pull-requests, etc.) Dan Bader has a nice article on what kind of things to write in a readme: https://dbader.org/blog/write-a-great-readme-for-your-github-project

That's a quite nice guide. I think most of the things where already in the readme. I added a section about how to contribute and restructured the readme.

> # Software Paper
>
>     * Summary is fine, but you misspelled Gaussian.

Ah yes, thank you.
