### 2021-03-08 AlexanderFabisch/gmr#25 @inakleinbottle
Sphinx is actually very easy to use, and all the complication is in the first setup. You have already done the hard work with writing the documentation strings. However, I'm not saying you must use sphinx, use whatever tool gets the job done.
