### 2021-03-13 14f2225
Minor fixes in readme
