### 2021-03-13 AlexanderFabisch/gmr#25 @AlexanderFabisch
> Sphinx is actually very easy to use, and all the complication is in the first setup. You have already done the hard work with writing the documentation strings. However, I'm not saying you must use sphinx, use whatever tool gets the job done.

I agree and I like sphinx, but I personally sometimes spend [too much time](https://rock-learning.github.io/pytransform3d/) fine-tuning the documentation.

I released 1.5, which should fix the errors in the unit tests and updated the paper with the your suggestions: https://github.com/openjournals/joss-reviews/issues/3054
