### 2021-03-14 AlexanderFabisch/gmr#25 @inakleinbottle
@AlexanderFabisch Your API documentation looks really good, I'm going to have a look at the tool you used. When you have some time, I'd look to flesh out some of the descriptions of the classes and functions. However, I'm satisfied that it meets the requirements for JOSS.

You've also done a stellar job sorting out the install requirements. I'm still having having problems installing the package in a fresh environment, which I think is because you import gmr in your setup script, which happens before Numpy is installed. You might be able to solve this problem by putting your version in a separate file within the gmr directory. Something like `__version__.py`, which you can import in `__init__.py` to get the version into the package number. This file could set the variable `__version__`. Then in your `setup.py` function script, you can `exec` this file instead of importing it, which will circumvent the Python import mechanics entirely and avoid the numpy imports from the main package. (You could use a function like this.)
```python3
def get_version():
    namespace = {}
    with open("gmr/__version__.py") as fp:
        exec(fp.read(), namespace)
    return namespace["__version__"]
```
I don't know if this is the best solution, but it will avoid the problems that occur at the moment. This might be overkill, but it will make sure.

Your instructions for contributing look good, I can tick this off now.

You've addressed the functionality comments that I made, so I'm happy to tick these off now.

I'm satisfied that you have addressed the differences between gmr and scikit-learn.

The only issue that is outstanding from my part is the installation process.
