### 2021-03-14 AlexanderFabisch/gmr#25 @AlexanderFabisch
> @AlexanderFabisch Your API documentation looks really good, I'm going to have a look at the tool you used. When you have some time, I'd look to flesh out some of the descriptions of the classes and functions.

Thanks, and yes, let me know if you have any suggestions.

>  I'm still having having problems installing the package in a fresh environment, which I think is because you import gmr in your setup script, which happens before Numpy is installed.

That's right. I didn't think about it. I just releast 1.5.1, which should fix it. At least it worked for me in a fresh docker container. I used sklearn's approach here:

* https://github.com/scikit-learn/scikit-learn/blob/main/setup.py#L32
* https://github.com/scikit-learn/scikit-learn/blob/main/sklearn/__init__.py#L59

Great feedback so far. Thank you!
