+++ paper/paper.bib
@@ -77,3 +77,10 @@ pages="1371--1394",
 isbn="978-3-540-30301-5",
 doi="10.1007/978-3-540-30301-5_60"
 }
+
+@techreport{Bishop1994,
+author = {Christopher M. Bishop},
+title = {Mixture Density Networks},
+institution = {Aston University},
+year = {1994}
+}
