### 2021-03-19 AlexanderFabisch/gmr#25 @inakleinbottle
@AlexanderFabisch I just tested the install again from PyPI and it seems to be working now. I must have got the previous version when I ran it last. Happy to tick this one off now.
