### 2021-03-19 AlexanderFabisch/gmr#25 @AlexanderFabisch
Thanks @inakleinbottle . Your review helped me a lot. In particular, the documentation is much better now.
