### 2021-04-03 68cf2f1
Add GMMRegression scikit-learn RegressorMixin
