### 2021-04-03 8a2e1e3
Add GMMRegression scikit-learn RegressorMixin
