### 2021-04-03 b2b9a54
Fixes numerical issue in GMM.condition
