### 2021-04-03 9ae6f36
Add unit test for numerical issue
