+++ gmr/tests/test_gmm.py
@@ -2,7 +2,8 @@ import sys
 import numpy as np
 from scipy.spatial.distance import pdist
 from gmr.utils import check_random_state
-from nose.tools import assert_equal, assert_less, assert_raises, assert_in, assert_false, assert_true
+from nose.tools import (assert_equal, assert_less, assert_raises, assert_in,
+                        assert_false, assert_true)
 from nose.plugins.skip import SkipTest
 from numpy.testing import assert_array_almost_equal
 try:
@@ -11,7 +12,8 @@ try:
 except ImportError:
     # Python 3
     from io import StringIO
-from gmr import GMM, MVN, plot_error_ellipses, kmeansplusplus_initialization, covariance_initialization
+from gmr import (GMM, MVN, plot_error_ellipses, kmeansplusplus_initialization,
+                 covariance_initialization)
 from test_mvn import AxisStub
 
 
@@ -440,3 +442,141 @@ def test_extract_mvns():
     assert_array_almost_equal(mvn0.mean, np.array([1, 2]))
     mvn1 = gmm.extract_mvn(1)
     assert_array_almost_equal(mvn1.mean, np.array([3, 4]))
+
+
+def test_condition_numerical_issue():
+    """Test for numerical issue in #27."""
+    covariances=np.array([
+        [[6.56478114e-03, 0.00000000e+00, 4.35794725e-01, 0.00000000e+00,
+          1.74768907e-03, -2.36645017e-02, 5.64492049e-01, -2.35229252e-02,
+          -1.87923556e-02, -9.39617778e-02, 2.44300622e-02, -6.63130035e-01,
+          3.85327366e-01, -4.16132516e-01],
+         [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00],
+         [4.35794725e-01, 0.00000000e+00, 1.33846496e+02, 0.00000000e+00,
+          5.36771200e-01, -5.15467320e+00, 7.84028133e+01, -4.98159335e+00,
+          -5.77173333e+00, -2.88586667e+01, 7.50325333e+00, -9.75587840e+01,
+          6.22656653e+01, -9.68002133e+01],
+         [0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00,
+          0.00000000e+00, 0.00000000e+00],
+         [1.74768907e-03, 0.00000000e+00, 5.36771200e-01, 0.00000000e+00,
+          2.15264000e-03, -2.06720400e-02, 3.14422667e-01, -1.99779293e-02,
+          -2.31466667e-02, -1.15733333e-01, 3.00906667e-02, -3.91244800e-01,
+          2.49707067e-01, -3.88202667e-01],
+         [-2.36645017e-02, 0.00000000e+00, -5.15467320e+00, 0.00000000e+00,
+          -2.06720400e-02, 5.04383573e-01, -4.56174933e+00, 1.93516556e-01,
[…]
