### 2021-04-03 8d18ffa
Apply oracle approximating shrinkage
