### 2021-04-03 714cd94
Refactor GMMRegression to a separate gmr/sklearn.py module
