### 2021-04-03 4334664
Refactor GMMRegression to a separate gmr/sklearn.py module
