### 2021-04-04 556e9ec
Refactor GaussianMixtureRegressor and add gmr/tests/test_sklearn.py test suite.
