+++ gmr/sklearn.py
@@ -1,26 +1,44 @@
 import numpy as np
-from sklearn.base import BaseEstimator, RegressorMixin, MultiOutputMixin
+
+try:
+    from sklearn.base import BaseEstimator, RegressorMixin, MultiOutputMixin
+    from sklearn.utils import check_X_y
+    from sklearn.utils.validation import check_is_fitted, check_array, FLOAT_DTYPES
+except ImportError:
+    raise ImportError("Install scikit-learn (e.g. pip install scikit-learn) to use this extension.")
 
 from .gmm import GMM
 
-class GMMRegression(MultiOutputMixin, RegressorMixin, BaseEstimator):
-    """Scikit-learn RegressorMixin for the GMM class.
+
+class GaussianMixtureRegressor(MultiOutputMixin, RegressorMixin, BaseEstimator):
+    """Gaussian mixture regression compatible to scikit-learn.
 
     Parameters
     ----------
-
-    n_components : int (> 0)
+    n_components : int
         Number of MVNs that compose the GMM.
+
+    priors : array, shape (n_components,), optional
+        Weights of the components.
+
+    means : array, shape (n_components, n_features), optional
+        Means of the components.
+
+    covariances : array, shape (n_components, n_features, n_features), optional
+        Covariances of the components.
+
+    verbose : int, optional (default: 0)
+        Verbosity level.
 
     random_state : int or RandomState, optional (default: global random state)
         If an integer is given, it fixes the seed. Defaults to the global numpy
         random number generator.
 
-    R_diff : float
+    R_diff : float, optional (default: 1e-4)
         Minimum allowed difference of responsibilities between successive
         EM iterations.
 
-    n_iter : int
+    n_iter : int, optional (default: 500)
         Maximum number of iterations.
 
     init_params : str, optional (default: 'random')
@@ -34,15 +52,10 @@ class GMMRegression(MultiOutputMixin, RegressorMixin, BaseEstimator):
         set based on the average distances of samples in each dimensions.
         This is computationally more expensive but often gives much
         better results.
-
-    Returns
-    -------
-    self : GMMRegression
-        This object.
     """
 
     def __init__(self, n_components, priors=None, means=None, covariances=None,
-                verbose=0, random_state=None, R_diff=1e-4, n_iter=500, init_params="random"):
+                 verbose=0, random_state=None, R_diff=1e-4, n_iter=500, init_params="random"):
         self.n_components = n_components
         self.priors = priors
         self.means = means
@@ -52,31 +65,30 @@ class GMMRegression(MultiOutputMixin, RegressorMixin, BaseEstimator):
         self.R_diff = R_diff
         self.n_iter = n_iter
         self.init_params = init_params
-    
-    def fit(self, X, y):
-        self.gmm = GMM(self.n_components, priors=self.priors, means=self.means, 
+
+        self.gmm_ = GMM(self.n_components, priors=self.priors, means=self.means,
                         covariances=self.covariances, verbose=self.verbose, random_state=self.random_state)
 
-        if y.ndim > 2:
-            raise ValueError("y must have at most two dimensions.")
-        elif y.ndim == 1:
+    def fit(self, X, y):
+        X, y = check_X_y(X, y, estimator=self.gmm_, dtype=FLOAT_DTYPES, multi_output=True)
+        if X.ndim == 1:
+            X = np.expand_dims(X, 1)
+        if y.ndim == 1:
             y = np.expand_dims(y, 1)
-        
-        if X.ndim > 2:
-            raise ValueError("y must have at most two dimensions.")
-        elif X.ndim == 1:
-            X = np.expand_dims(X, 1)
 
[…]
