### 2021-04-04 b5f1e06
Refactor GaussianMixtureRegressor and add gmr/tests/test_sklearn.py test suite.
