+++ gmr/tests/test_gmm.py
@@ -13,7 +13,7 @@ except ImportError:
     # Python 3
     from io import StringIO
 from gmr import (GMM, MVN, plot_error_ellipses, kmeansplusplus_initialization,
-                 covariance_initialization, GMMRegression)
+                 covariance_initialization)
 from test_mvn import AxisStub
 
 
@@ -263,12 +263,6 @@ def test_regression_with_2d_input():
     y = np.vstack((y1, y2)) + noise
     samples = np.hstack((x, x[::-1], y))
 
-    gmm = GMMRegression(n_components=2, random_state=random_state)
-    gmm.fit(np.hstack((x, x[::-1])), y)
-
-    pred = gmm.predict(np.hstack((x, x[::-1])))
-    mse = np.sum((y - pred) ** 2) / n_samples
-
 
 def test_regression_without_noise():
     """Test regression without noise."""
@@ -291,17 +285,6 @@ def test_regression_without_noise():
     mse = np.sum((y - pred) ** 2) / n_samples
     assert_less(mse, 0.01)
 
-    random_state = check_random_state(0)
-
-    gmm = GMMRegression(n_components=2, random_state=random_state)
-    gmm.fit(x, y)
-    assert_array_almost_equal(gmm.gmm.priors, 0.5 * np.ones(2), decimal=2)
-    assert_array_almost_equal(gmm.gmm.means[0], np.array([1.5, 2.5]), decimal=2)
-    assert_array_almost_equal(gmm.gmm.means[1], np.array([0.5, 2.5]), decimal=1)
-
-    pred = gmm.predict(x)
-    mse = np.sum((y - pred) ** 2) / n_samples
-    assert_less(mse, 0.01)
 
 def test_plot():
     """Test plot of GMM."""
