### 2021-04-04 68f7e84
Document prior computation in GMM.condition
