+++ gmr/__init__.py
@@ -5,7 +5,7 @@ gmr
 Gaussian Mixture Models (GMMs) for clustering and regression in Python.
 """
 
-__version__ = "1.5.1"
+__version__ = "1.6.dev"
 
 try:
     # Idea from sklearn:
