### 2021-04-14 f8ce55f
Refactor gmr/sklearn.py
