### 2021-04-14 de80442
Refactor gmr/sklearn.py
