### 2021-04-14 78585bc
PEP8 and document attributes
