+++ gmr/sklearn.py
@@ -3,9 +3,12 @@ import numpy as np
 try:
     from sklearn.base import BaseEstimator, RegressorMixin, MultiOutputMixin
     from sklearn.utils import check_X_y
-    from sklearn.utils.validation import check_is_fitted, check_array, FLOAT_DTYPES
+    from sklearn.utils.validation import (check_is_fitted, check_array,
+                                          FLOAT_DTYPES)
 except ImportError:
-    raise ImportError("Install scikit-learn (e.g. pip install scikit-learn) to use this extension.")
+    raise ImportError(
+        "Install scikit-learn (e.g. pip install scikit-learn) to use this "
+        "extension.")
 
 from .gmm import GMM
 
@@ -52,10 +55,19 @@ class GaussianMixtureRegressor(MultiOutputMixin, RegressorMixin, BaseEstimator):
         set based on the average distances of samples in each dimensions.
         This is computationally more expensive but often gives much
         better results.
+
+    Attributes
+    ----------
+    gmm_ : GMM
+        Underlying GMM object
+
+    indices_ : array, shape (n_features,)
+        Indices of inputs
     """
 
     def __init__(self, n_components, priors=None, means=None, covariances=None,
-                 verbose=0, random_state=None, R_diff=1e-4, n_iter=500, init_params="random"):
+                 verbose=0, random_state=None, R_diff=1e-4, n_iter=500,
+                 init_params="random"):
         self.n_components = n_components
         self.priors = priors
         self.means = means
@@ -67,17 +79,21 @@ class GaussianMixtureRegressor(MultiOutputMixin, RegressorMixin, BaseEstimator):
         self.init_params = init_params
 
     def fit(self, X, y):
-        self.gmm_ = GMM(self.n_components, priors=self.priors, means=self.means,
-                        covariances=self.covariances, verbose=self.verbose, random_state=self.random_state)
+        self.gmm_ = GMM(
+            self.n_components, priors=self.priors, means=self.means,
+            covariances=self.covariances, verbose=self.verbose,
+            random_state=self.random_state)
 
-        X, y = check_X_y(X, y, estimator=self.gmm_, dtype=FLOAT_DTYPES, multi_output=True)
+        X, y = check_X_y(X, y, estimator=self.gmm_, dtype=FLOAT_DTYPES,
+                         multi_output=True)
         if y.ndim == 1:
             y = np.expand_dims(y, 1)
 
         self.indices_ = np.arange(X.shape[1])
 
-        self.gmm_.from_samples(np.hstack((X, y)),
-                               R_diff=self.R_diff, n_iter=self.n_iter, init_params=self.init_params)
+        self.gmm_.from_samples(
+            np.hstack((X, y)), R_diff=self.R_diff, n_iter=self.n_iter,
+            init_params=self.init_params)
         return self
 
     def predict(self, X):
