### 2021-04-14 3c18fd9
Document model selection with regressor
