### 2021-04-14 6cdc486
Add assertion in test
