### 2021-04-14 e62c19a
Skip tests in case sklearn is not available
