### 2021-04-14 c877e58
Install sklearn for unit tests
