### 2021-04-15 138c30d
Fix numerical issue in expectation step
