### 2021-04-15 46db101
Fix numerical issue in expectation step
