### 2021-04-19 10e24a3
Fast prediction of conditional mean
