### 2021-04-19 2f042f8
Fast prediction of conditional mean
