### 2021-04-19 cb4b69b
Refactor safe prob. density
