### 2021-04-19 1c8c0a1
More refactoring
