### 2021-05-05 c7f7ef2
Correct formulation in JOSS paper
