### 2021-05-28 #3054 @soodoku
Well done, @AlexanderFabisch!

I have heartache about original implementations of a bunch of fundamental things, starting with something as trivial as mahalanobis distance. It is generally a bad idea as there are always lots of edge cases that the Python codebase would generally go over.

I cannot do a line-by-line analysis of whether core functionality has been correctly implemented in a reasonable amount of time. If we are not fussed by that, everything else seems above par to me.
