### 2021-05-28 #3054 @AlexanderFabisch
> Well done, @AlexanderFabisch!

Thank you.

> I have heartache about original implementations of a bunch of fundamental things, starting with something as trivial as mahalanobis distance. It is generally a bad idea as there are always lots of edge cases that the Python codebase would generally go over.

That's a very good point actually. I will go through the code again and see what I can replace with numpy / scipy functions. For example, there is [scipy.spatial.distance.mahalanobis](https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.distance.mahalanobis.html), which is actually almost identical to my implementation, but it's still a good idea to use it.

@mikldk I would address this issue raised by @soodoku , release a new version of the package, and check the paper. I will let you know when I'm done.
