### 2021-05-28 6b84184
Use mahalanobis distance from scipy
