### 2021-05-28 7ab7fc6
Check if MVN is initialized in mahalanobis
