### 2021-05-28 970c815
Use pinvh from scipy
