### 2021-05-28 #3054 @AlexanderFabisch
OK, I'm done with the last changes.

* [Here](https://github.com/AlexanderFabisch/gmr/pull/35) is the pull request in which I try to replace some basic functions by scipy / numpy. Here are my comments:
    * [scipy.stats.multivariate_normal.pdf](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.multivariate_normal.html#scipy.stats.multivariate_normal) is not robust enough to handle singular covariances. In general `gmr` takes care of handling many numerical pitfalls accurately (often in a different way than sklearn's implementation of GMMs by the way) and improved a lot in recent updates. This is a crucial aspect for a robust GMR implementation and it is often not directly supported by scipy and numpy.
    * pinvh is part of scipy (now?), it has been replaced
    * scipy.spatial.distance.mahalanobis will be used
    * I'd like to replace more of the basic functions by numpy / scipy functions, I just couldn't find more at the moment.
* I made a last change to the paper [here](https://github.com/AlexanderFabisch/gmr/pull/36/files).
* I released `gmr` 1.6.

Although I made a GitHub release and the Zenodo webhook worked for the last releases, there is now new DOI in Zenodo for release 1.6. However, [10.5281/zenodo.4449631](https://doi.org/10.5281/zenodo.4449631) should be the DOI of the latest release on Zenodo.

> Please make a tagged release and archive (e.g. with Zenodo) as described here, and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.

I hope it is OK when the author list on Zenodo and on the paper are different. João Felipe Santos is listed as a co-author at Zenodo. That is OK because he contributed to the repository (Python 3 compatibility), but I would like to be the only author of the paper.
