### 2021-05-31 #3054 @arfon
> Felipe Santos is listed as a co-author at Zenodo. That is OK because he contributed to the repository (Python 3 compatibility), but I would like to be the only author of the paper.

@AlexanderFabisch - have you contacted Felipe about this and confirmed they are happy with this?
