### 2021-05-31 #3054 @AlexanderFabisch
We are talking about [2 commits](https://github.com/AlexanderFabisch/gmr/graphs/contributors). The current version of Zenodo also doesn't list him as an author. Should I contact him? If yes, then I'll try it now.
