### 2021-05-31 #3054 @arfon
> We are talking about 2 commits. The current version at Zenodo also doesn't list him as an author. Should I contact him? If yes, then I'll try it now.

I think this is probably OK, especially as the changes are very minimal.
