### 2021-06-01 #3054 @AlexanderFabisch
@mikldk I totally forgot to add acknowledgments to the paper unfortunately. I did that [now](https://github.com/AlexanderFabisch/gmr/commit/3653977051a15e0ff0bacdc1d4989c70c4905eb9). Sorry for the chaos. Would that mean that I have to make a new release to Zenodo? I corrected the title at Zenodo.
