### 2021-06-01 #3054 @AlexanderFabisch
Mmh... yes. I am sorry. Zenodo is a bit weird. I don't know how it determines the authors of the software. :)

Marcelo Albuquerque contributed a sklearn-compatible interface some weeks ago. That was after I submitted this paper so he doesn't appear in the statistics above.

So I would have a general question about the authorship policy. How should we handle the authorship of the paper and the software? As I understand it, the authors of a JOSS paper are not necessarily authors of the paper itself but authors of the software. [Here](https://github.com/AlexanderFabisch/gmr/graphs/contributors) you can see the amount of contribution of each author. Marcelo Albuquerque actually contributed although the percentage is small and the contributions were after submission of the paper. Does this already qualify for authorship? If yes, then give me the time to ask him. In this case I would agree that it could make sense to include him.

Again, sorry for the chaos. I could have raised the issue yesterday when we were discussing this issue.
