### 2021-06-01 AlexanderFabisch/gmr#37 @AlexanderFabisch
JOSS paper authorship
I recently submitted a paper to the [journal of open source software](https://joss.theoj.org/). Submitted papers are typically tightly coupled with Zenodo releases. Here is the latest release of GMR: https://zenodo.org/record/4889867 and here is the paper review issue: https://github.com/openjournals/joss-reviews/issues/3054

In addition to myself, Zenodo also lists you, @jfsantos and @mralbu, as authors. That is why I would have to agree with you both on whether you want to become author of the JOSS paper as well. Both author lists should be as similar as possible.

@jfsantos The amount of lines of code that you contributed and that survived in the latest version of gmr is [8](https://github.com/openjournals/joss-reviews/issues/3054#issuecomment-783269820) so that I suggest that it would not make sense to include you as an author if that is OK. Do you want to stay author of the Zenodo release or should I remove your name here? I wouldn't have any issue with any of the two options.

@mralbu You contributed the sklearn interface and I adopted your idea for doing faster batch mean predictions. So I would ask you whether you want to become co-author of the paper. Have a look at the latest article proof here: https://github.com/openjournals/joss-reviews/issues/3054#issuecomment-852008730 . The current state of the review is that both reviewers accepted the paper and we are currently only discussing the authorship details.
