### 2021-06-01 AlexanderFabisch/gmr#37 @jfsantos
Hi Alexander,

the changes I did were minimal and setup-related, and do not warrant authorship for sure. Feel free to remove my name from the list of authors.
