### 2021-06-01 AlexanderFabisch/gmr#37 @AlexanderFabisch
I read that as I should remove you as author from the Zenodo release and from the JOSS paper. If that is a misinterpretation, please let be know.
