### 2021-06-01 AlexanderFabisch/gmr#37 @mralbu
Hi @AlexanderFabisch,
Thanks for the invitation. My contributions were not very significative, and I didn't contribute with the paper, so it would be ok not to be included as author as well.
