### 2021-06-01 AlexanderFabisch/gmr#37 @AlexanderFabisch
@mralbu Thanks for your feedback.

Also thank you both for your contributions, which helped to improve the library. :)
