### 2021-06-01 #3054 @AlexanderFabisch
@mikldk We came to a conclusion [here](https://github.com/AlexanderFabisch/gmr/issues/37): I will be the only author of the JOSS paper. The author list of the Zenodo release is still different though.
