### 2021-06-05 #3054 @Kevin-Mattheus-Moerman
@AlexanderFabisch I'll be helping to get your paper ready for publication in JOSS. Below are some remaining issues.

- [ ] Please amend the Zenodo archive meta-data to match that of the paper. In particular the author list is not the same.
- [ ]  Can you confirm `sklearn` is correct in `..while sklearn has...`, or should this be skikit-learn. If it is a different package can you cite it?
- [ ] Both `multimodal` and `multi-modal` occur in the text, please choose one and correct throughout to be consistent.
