### 2021-06-05 6522f83
JOSS: make multimodal consistent
