+++ paper/paper.md
@@ -39,7 +39,7 @@ is not available in scikit-learn and, thus, will be provided by gmr.
 Note that while sklearn has the function GaussianMixture.predict,
 it does not perform regression. This function computes the index of the
 Gaussian for which the inputs have the highest probability. Furthermore,
-multi-modal regression often requires a more complicated interface to extract
+multimodal regression often requires a more complicated interface to extract
 not just the mean but also individual Gaussians or sample from the predicted
 distribution.
 
