### 2021-06-05 e50890d
JOSS: use scikit-learn consistently
