### 2021-06-05 #3054 @AlexanderFabisch
@Kevin-Mattheus-Moerman Thanks. I addressed these points:

>     * [ ]  Please amend the Zenodo archive meta-data to match that of the paper. In particular the author list is not the same.

OK, I give up, for the corresponding Zenodo release the author lists are the same now.

>     * [ ]   Can you confirm `sklearn` is correct in `..while sklearn has...`, or should this be skikit-learn. If it is a different package can you cite it?

I replaced sklearn by scikit-learn.

>     * [ ]  Both `multimodal` and `multi-modal` occur in the text, please choose one and correct throughout to be consistent.

I use "multimodal" consistently now.

Should I make a new Zenodo release with these additional changes?
