### 2021-06-06 #3054 @AlexanderFabisch
> One recommendation I have is to also add your ORCID profile to the author listed on Zenodo.

The ORCID profile already should have been linked on Zenodo.

> @AlexanderFabisch congratulations on your published paper!

Thanks!

Also thanks to you @inakleinbottle, @soodoku, and @mikldk for reviewing and editing!
